+++ toys-core/test/test_loader.rb
@@ -45,32 +45,32 @@ describe Toys::Loader do
 
     it "finds a tool directly defined in a config file" do
       tool = loader.lookup(["tool-1"])
-      assert_equal("file tool-1 short description", tool.effective_desc)
-      assert_equal("file tool-1 long description", tool.effective_long_desc)
+      assert_equal(["file tool-1 short description"], tool.effective_desc)
+      assert_equal(["file tool-1 long description"], tool.effective_long_desc)
     end
 
     it "finds a subtool directly defined in a config file" do
       tool = loader.lookup(["group-1", "tool-1-1"])
-      assert_equal("file tool-1-1 short description", tool.effective_desc)
-      assert_equal("file tool-1-1 long description", tool.effective_long_desc)
+      assert_equal(["file tool-1-1 short description"], tool.effective_desc)
+      assert_equal(["file tool-1-1 long description"], tool.effective_long_desc)
       assert_equal(["group-1", "tool-1-1"], tool.full_name)
     end
 
     it "finds a group directly defined in a config file" do
       tool = loader.lookup(["group-1"])
-      assert_equal("file group-1 short description", tool.effective_desc)
+      assert_equal(["file group-1 short description"], tool.effective_desc)
       assert_equal(["group-1"], tool.full_name)
     end
 
     it "finds a tool defined in a file in a config directory" do
       tool = loader.lookup(["tool-2"])
-      assert_equal("directory tool-2 short description", tool.effective_desc)
-      assert_equal("directory tool-2 long description", tool.effective_long_desc)
+      assert_equal(["directory tool-2 short description"], tool.effective_desc)
+      assert_equal(["directory tool-2 long description"], tool.effective_long_desc)
     end
 
     it "finds the nearest group directly defined if a query doesn't match" do
       tool = loader.lookup(["group-1", "tool-blah"])
-      assert_equal("file group-1 short description", tool.effective_desc)
+      assert_equal(["file group-1 short description"], tool.effective_desc)
       assert_equal(["group-1"], tool.full_name)
     end
 
@@ -88,14 +88,14 @@ describe Toys::Loader do
 
     it "finds a tool directly defined" do
       tool = loader.lookup(["tool-1"])
-      assert_equal("normal tool-1 short description", tool.effective_desc)
-      assert_equal("normal tool-1 long description", tool.effective_long_desc)
+      assert_equal(["normal tool-1 short description"], tool.effective_desc)
+      assert_equal(["normal tool-1 long description"], tool.effective_long_desc)
     end
 
     it "finds a subtool directly defined" do
       tool = loader.lookup(["group-1", "tool-1-3"])
-      assert_equal("normal tool-1-3 short description", tool.effective_desc)
-      assert_equal("normal tool-1-3 long description", tool.effective_long_desc)
+      assert_equal(["normal tool-1-3 short description"], tool.effective_desc)
+      assert_equal(["normal tool-1-3 long description"], tool.effective_long_desc)
       assert_equal(["group-1", "tool-1-3"], tool.full_name)
     end
 
@@ -142,8 +142,8 @@ describe Toys::Loader do
 
     it "allows loading if the collision isn't actually traversed" do
       tool = loader.lookup(["tool-2"])
-      assert_equal("index tool-2 short description", tool.effective_desc)
-      assert_equal("index tool-2 long description", tool.effective_long_desc)
+      assert_equal(["index tool-2 short description"], tool.effective_desc)
+      assert_equal(["index tool-2 long description"], tool.effective_long_desc)
     end
 
     it "reports error if a tool is defined multiple times" do
@@ -160,8 +160,8 @@ describe Toys::Loader do
       loader.add_path(File.join(cases_dir, "normal-file-hierarchy"))
 
       tool = loader.lookup(["tool-1"])
-      assert_equal("file tool-1 short description", tool.effective_desc)
-      assert_equal("file tool-1 long description", tool.effective_long_desc)
+      assert_equal(["file tool-1 short description"], tool.effective_desc)
+      assert_equal(["file tool-1 long description"], tool.effective_long_desc)
     end
 
     it "honors the high-priority flag" do
@@ -170,8 +170,8 @@ describe Toys::Loader do
       loader.add_path(File.join(cases_dir, "normal-file-hierarchy"), high_priority: true)
 
       tool = loader.lookup(["tool-1"])
-      assert_equal("normal tool-1 short description", tool.effective_desc)
-      assert_equal("normal tool-1 long description", tool.effective_long_desc)
+      assert_equal(["normal tool-1 short description"], tool.effective_desc)
+      assert_equal(["normal tool-1 long description"], tool.effective_long_desc)
     end
   end
 
@@ -182,20 +182,20 @@ describe Toys::Loader do
 
     it "gets an item from a root-level directory include" do
       tool = loader.lookup(["tool-2"])
-      assert_equal("directory tool-2 short description", tool.effective_desc)
-      assert_equal("directory tool-2 long description", tool.effective_long_desc)
+      assert_equal(["directory tool-2 short description"], tool.effective_desc)
+      assert_equal(["directory tool-2 long description"], tool.effective_long_desc)
     end
 
     it "gets an item from a root-level file include" do
       tool = loader.lookup(["group-1", "tool-1-1"])
-      assert_equal("file tool-1-1 short description", tool.effective_desc)
-      assert_equal("file tool-1-1 long description", tool.effective_long_desc)
+      assert_equal(["file tool-1-1 short description"], tool.effective_desc)
+      assert_equal(["file tool-1-1 long description"], tool.effective_long_desc)
     end
 
     it "gets an item from non-root-level include" do
       tool = loader.lookup(["group-0", "group-1", "tool-1-1"])
-      assert_equal("normal tool-1-1 short description", tool.effective_desc)
-      assert_equal("normal tool-1-1 long description", tool.effective_long_desc)
+      assert_equal(["normal tool-1-1 short description"], tool.effective_desc)
+      assert_equal(["normal tool-1-1 long description"], tool.effective_long_desc)
     end
 
     it "does not load an include if not needed" do