+++ toys/CHANGELOG.md
@@ -3,8 +3,10 @@
 ### 0.3.3 / TBD
 
 * CHANGED: Renamed file_utils helper to fileutils.
+* FIXED: Documentation strings that begin with "--" no longer cause problems.
 * ADDED: Highline helper
 * ADDED: Spinner helper
+* IMPROVED: Descriptions can have multiple lines
 
 ### 0.3.2 / 2018-05-07